Ensure that two double attributes inserted with the same long value create a single concept (#5808)
## What is the goal of this PR?

Previously, given: `define length sub attribute, value double`, when you did:

```
insert
$x 2 isa length;
$y 2 isa length;
```

**two** concepts were created instead of one. Now, only a single concept is created.

## What are the changes implemented in this PR?

The value of a `double` attribute needs to be normalised in order to accurately check for equality. This was handled effectively by the `AttributeValueConverter`, which provides a mechanism for taking in an attribute value and its type, and returning the value in a normalised form that ensures, for example, that `2`, `2.0` and `2.000` are all converted to the same output value because they represent the same value.

However, the `AttributeValueConverter` was only used in a small fraction of all the instances where we need to generate the "unique index" used by JanusGraph for deduplication. The method used to generate this index is at a very high level, in `Schema.java`, and cannot be easily replaced. Therefore, this PR aims to address the issue by ensuring that we pass in the correct `value` parameter to `Schema.java` every time, by consistently using `AttributeValueConverter.tryConvert` to convert the attribute value to normal form before it is used.

Also, it has a try-catch block in it which ensures we consistently throw the same error if the conversion could not be performed (e.g: attempting to parse "banana" as a `long`).